+++ package.json
@@ -38,6 +38,7 @@
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-image": "^4.1.0",
     "react-intersection-observer": "^9.5.2",
     "react-moment": "^1.1.3",
     "react-number-format": "^5.3.1",
